Add network on/off toggle. Accumulate/sync.
Add network on/off toggle to UI.

When the network is off, we execute the ops locally and display
them. However, we also record a snapshot of the state at the
point the network connection was lost and we record all ops
into a single atomic transaction. When the network connection
is restored, the state is rolled back to the snapshot and the
atomic transaction is sent to the server for processing. The
use may or may not lose some of its work.